requirements.txt
Django==2.2.13
# postgreSQLドライバ
psycopg2==2.7.6
psycopg2-binary==2.7.7
# reactとの連携に必要
django-cors-headers==2.5.2
# herokuに必要
gunicorn
django-heroku
whitenoise
# DRFに必要
markdown
django-filter
djangorestframework==3.9.2
# メール認証付きユーザー登録機能に必要
django-allauth